requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.3.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
